requirements.txt
torch>=1.7.0
torchvision>=0.8.0
numpy>=1.19.2
Pillow>=8.0.0
tqdm>=4.50.0
transformers>=4.3.0
chardet>=4.0.0
